Revert "Add ability to filter params and default to passwords"
This reverts commit d3a2557dc20d2cf77a9010522ec40214fa41bb64.

This change was simply made in the wrong place because I am a moron. I
made changes to the LogSubscriber; which hooks into ActionController
anyway... so already does all this bloody stuff.

The change is needed in the LogFormatter for non-rails apps.

Reverting these changes as they add unnecessary processing and codejunk.

Bad Ed, you get a kick in the shin.